Migrate export_saved_objects data source to envelope pattern
